Modify signals.pm and slots.pm to allow the creation of overloaded signals and slots, aka ones with the same name, but accepting different arguments.  All slots with the same name call one perl sub.
There was a bug with metaObject creation.  The code wasn't using a moduleId,
but was hard-coded to use qtcore_Smoke.  This was resulting in the metaObject's
superClass() to be wrong for the KDE classes.  Fixed.

There was also a bug where a method defined in one smoke object returned a type
whose class was defined in a different smoke object.  This would cause the
resolve_classname function to not be able to find the perl classname.  Fixed.

svn path=/trunk/playground/bindings/perlqt4/; revision=1109914