Add support for Qt::AbstractXmlNodeModel::createIndex and Qt::XmlNodeModelIndex::internalPointer
In puic4, always check for known classes.  The .ui file could have a Qt header specified explicitly, and we need to resolve that to a PerlQt module name.

Now that the caller() method can be used to correctly report what the current file and line is, use it in more error messages.

Add the xmlpatterns/qobjectxmlmodel example.

svn path=/trunk/KDE/kdebindings/perl/; revision=1184899