Initial changes for Windows support.
Here are the problems:
* PerlQt .dlls and Qt dlls use the same name ("QtCore4.dll", "QtGui4.dll").
So the PerlQt one gets loaded, but the Qt one doesn't.  The only way to
get around this is to rename the .dll.  But I don't want to rename all the
Perl modules, because that would break current code.  So you'd end up with
the .pm files having a different name from the .dll files.

* ExtUtils::MakeMaker uses a separate file to list what functions should
be visible to outside code, specifically marking the boot functions as
visible.  PerlQt uses CMake, which doesn't appear to support this,
expecting the visibility attributes to be set in the code.  But Perl uses
macros to declare the XS functions, so in order to set the visibility in
the code, I have to redefine one of the Perl macros to inject
"__declspec(dllexport)".

I'm looking at
http://social.msdn.microsoft.com/Forums/en-US/vcgeneral/thread/b3eaa07f-7f92-4693-8aa1-b8fee0b92d2f/
to see if using a .manifest can get around having to rename QtCore4.dll