Add QMultiMap handler for the phonon qmusicplayer example.
It's a bit odd, because in C++ they save the result of MediaObject::metaData in
a QMap, even though the method returns a QMultiMap.  This works in C++ because
QMultiMap is a subclass, but there's no convenient way to do that in Perl.  So
it is left to the user of the method to get the first element of the array of
each hash key.