Each perlqt_module can now supply its own MarshallSlotReturnValueFn.
This allows the individual modules to be responsible for any special types that
need to be handled when marshalling the return value of a slot call.  This is
being used to keep the QtCore module from needing to link directly to the
QtDBus one (which needs special behavior for the QDBusVariant type).  Windows
and OS X don't have the DBus module, so we need to be able to build without
linking to QtDBus.