Changes to support perl 5.18.0.
Perl no longer calls the AUTOLOAD method for DESTROY.  So that has to be defined explicitly.
It also didn't like bless \shift, $package, so I've changed that as well.
defined @array is deprecated, update that code.
Also remove spurious breakpoints in example code.
Also, CMake seems to have changed, so the prove "macro" is now a function, because functions support actual scope on cmake variables.