Use a copy of the Qt::AutoLoad::AUTOLOAD variable.
This appears to be a change in Perl's behavior.  The way the operator
overloading in PerlQt works is that it first tries to find an operator method
on the class itself, and then next it tries to find one in the so-called
QGlobalSpace, which is a place defined by the smoke library for global Qt
functions.  Perl passes the underlying XS code the full package and function
being called, which PerlQt splits into 2 strings, one for the package name, and
one for the method name.  PerlQt null-terminates the package name string, and
in previous versions of Perl, this modification did not affect the source
$AUTOLOAD variable.  In Perl 5.20.0, it does update the Perl variable, and then
causes confusion down the line.

https://bugzilla.redhat.com/show_bug.cgi?id=1136340
https://bugs.kde.org/show_bug.cgi?id=340518