Fix cmake dependency on optional smoke components.
PerlQt depends on the smoke wrapper library for a component, not the source
library itself.  So if, for instance, Qwt is installed, but the smoke lib is
not present, perl bindings to qwt cannot be built.  This change updates the
dependency to be on the smoke qwt library instead.

Fixes issue 63.