Maintain order of methodIds when running them through uniqMethods()
The order of the methods defines their score.  The methodId on the front
of the list is the best match for the given call.  If we only store them in
a hash and return the values, the original order is lost.  Instead, store
the methods in an array, and their signature in the hash.  If the signature
is already in the hash, don't add it to the array.