Add handler for void**.  This is needed for qt_metacall().
In order for qt_metacall() to be seen on subclasses of QObject that define
their own QMetaObject (when they define their own signals, slots,
properties, etc), qt_metacall must be defined as a method in the
appropriate perl package.  Then when the qt_metacall() virtual method is
called, the SmokePerlBinding::callMethod() will invoke the Perl-defined
qt_metacall().  This requires that the void** type be sent through as a
perl variable, which the xsub can then read to get the void** out.