can() always returns a subclass of SmokePerl::Method.
QtCore defines 2 subclasses of SmokePerl::Method: PerlQt5::QtCore::Slot and
PerlQt5::QtCore::Signal.  The correct subclass is found by the
SmokePerl::Method constructor, effectively making it a factory method.

SmokePerl::Method remembers which object it was called with, and passes
that as the first argument when it is called.  It overrides the subroutine
dereference operator, so it can be used identically to a coderef, but can
also have methods attached to it.