Method subclasses don't store a coderef anymore.
If the method found by can() is XS_AUTOLOAD, then the string value of the
internal $AUTOLOAD stored on the cv could change in the future.  Since
the method instances store the method name anyway, just redo the lookup
at call time, to force $AUTOLOAD to be set correctly again.