Override QObject::can so that it can return Signal instances for private signals
If the last argument for a signal is QPrivateSignal
(see https://woboq.com/blog/how-qt-signals-slots-work-part2-qt5.html), then
that signal is only emittable by the class that declared it.  It will not
appear in the list of methods in the smoke module.

In this case, we look at the list of methods that the QMetaObject tells us
about.  The private signal still cannot be called, but it can be connected
to.  The objective is for code like this to work:

    my $app = QApplication->new(\@ARGV);
    $app->can('aboutToQuit')->connect(\&beforeExitHandler);

even though

    $app->can('aboutToQuit')->();

would fail.