Avoid conflicts between Qt's and Perl's use of environment variables.
From http://www.eprints.org/tech.php/13984.html:
When Perl is complied with PERL_USE_SAFE_PUTENV, or PL_use_safe_putenv is
set:
If an environment variable is unset or reset, perl frees the previous
environment string and allocates a new one using its own memory
management routines if applicable. If the environment is extended by a
new variable, perl reallocates the environ array using its own memory
management. This works fine if the environment is only accessed from
perl but it will fail if a C-extension also sets environment variables.
A variable that is set by a C-extension may not be freed by perl later
because perl cannot know, wether the extension still uses the string or
not. Moreover reallocations of the environ array may not be mixed
between perl and stdlib memory management.

Unfortunately the Qt libraries are one of those
C-extensions that modify the environment. On a perl version compiled
without PERL_USE_SAFE_PUTENV (like the one that ships with RHEL5), this
can cause invalid frees at exit.