Track valid state separately from an object's pointer.
In order to remove an Object from the pointer map, that Object's value
pointer has to be correct.  If value is set to nullptr to indicate that the
object is invalid, then when ObjectMap::remove() is called, it doesn't know
the key for that object in the map.  This means that invalidate() has to be
a public method so that others can call it, and a separate field has to be
maintained to determine if an object is valid or not.  The main way that
objects become invalid is when their parent is deleted.