Load a .dll prefixed with "Perl" on Windows.
On Windows, there could be a namespace collision between the DLLs that the
bindings wrap, and the DLLs that are the Perl extensions.  To work around
this, prefix the Perl extensions with the text "Perl".  This requires a few
hoops to be jumped through though.  The module boot function name is
expected to match the Perl module name.  So on Windows, we define a second
boot function, that forwards the call to the real boot function.