Use the B module to more easily get the refcount of an object
It seems that dereferencing a hash, and passing that to
Devel::Peek::SvREFCNT doesn't work anymore. The thing that is sent to
the method is the first key in the hash, so you get the refcount of that
instead of the refcount of the hash itself. The B module provides the
support needed for these tests with a more direct syntax anyway.